+++ frontend/app/src/main/java/de/amosproj3/ziofa/ZiofaApplication.kt
@@ -7,14 +7,14 @@ package de.amosproj3.ziofa
 import android.app.Application
 import android.content.Context
 import android.content.pm.PackageManager
-import de.amosproj3.ziofa.api.BackendConfigurationAccess
-import de.amosproj3.ziofa.api.DataStreamProvider
-import de.amosproj3.ziofa.api.LocalConfigurationAccess
-import de.amosproj3.ziofa.api.RunningComponentsAccess
-import de.amosproj3.ziofa.bl.ConfigurationManager
-import de.amosproj3.ziofa.bl.DataStreamManager
-import de.amosproj3.ziofa.bl.PackageInformationProvider
-import de.amosproj3.ziofa.bl.RunningComponentsProvider
+import de.amosproj3.ziofa.api.configuration.BackendConfigurationAccess
+import de.amosproj3.ziofa.api.events.DataStreamProvider
+import de.amosproj3.ziofa.api.configuration.LocalConfigurationAccess
+import de.amosproj3.ziofa.api.processes.RunningComponentsAccess
+import de.amosproj3.ziofa.bl.configuration.ConfigurationManager
+import de.amosproj3.ziofa.bl.events.DataStreamManager
+import de.amosproj3.ziofa.bl.processes.PackageInformationProvider
+import de.amosproj3.ziofa.bl.processes.RunningComponentsProvider
 import de.amosproj3.ziofa.client.ClientFactory
 import de.amosproj3.ziofa.client.RustClientFactory
 import de.amosproj3.ziofa.ui.configuration.ConfigurationViewModel
