+++ frontend/app/src/main/java/de/amosproj3/ziofa/api/configuration/ConfigurationUpdate.kt
@@ -3,7 +3,7 @@
 //
 // SPDX-License-Identifier: MIT
 
-package de.amosproj3.ziofa.api
+package de.amosproj3.ziofa.api.configuration
 
 import de.amosproj3.ziofa.client.Configuration
