+++ frontend/app/src/main/java/de/amosproj3/ziofa/api/processes/RunningComponentsAccess.kt
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: MIT
 
-package de.amosproj3.ziofa.api
+package de.amosproj3.ziofa.api.processes
 
 import kotlinx.coroutines.flow.Flow
