+++ frontend/app/src/main/java/de/amosproj3/ziofa/ui/visualization/utils/VisualizationHelpers.kt
@@ -4,7 +4,7 @@
 
 package de.amosproj3.ziofa.ui.visualization.utils
 
-import de.amosproj3.ziofa.api.RunningComponent
+import de.amosproj3.ziofa.api.processes.RunningComponent
 import de.amosproj3.ziofa.ui.shared.toReadableString
 import de.amosproj3.ziofa.ui.visualization.data.DropdownOption
